整理目录和文件结构，删除了一些过时的文件（diesel 的 schema）
