+++ docs/错误处理.md
@@ -24,7 +24,7 @@ pub struct ServerError {
 `ServerError` 转换成 `ErrorResponse` 的过程中，`ServerError` 序列化为 JSON 字符串存于响应主体里，如：
 
 ```json
-{code: 1, msg: "Internal server error."}
+{code: 1, msg: "Internal services error."}
 ```
 
 *注：项目中约定服务器尽可能返回 HTTP 200 OK，然后在响应主体里说明错误代码。*
