+++ src/main.rs
@@ -1,4 +1,4 @@
-#![allow(non_snake_case)]
+#![allow(dead_code)]
 
 extern crate chrono;
 #[macro_use]
@@ -7,24 +7,16 @@ extern crate failure;
 extern crate lazy_static;
 #[macro_use]
 extern crate num_derive;
-#[macro_use]
 extern crate log;
 
-pub mod attachment;
-pub mod config;
-pub mod error;
-pub mod freshman;
+mod config;
+mod error;
 mod jwt;
-mod motto;
-pub mod sale;
-pub mod schema;
-pub mod server;
-pub mod sign;
-pub mod user;
-pub mod wechat;
+mod models;
+mod services;
 
 // Import main function.
-use crate::server::server_main;
+use crate::services::server_main;
 
 fn main() {
     server_main().unwrap_or_else(|e| {
