+++ src/models.rs
@@ -1,0 +1,6 @@
+pub mod attachment;
+pub mod freshman;
+pub mod motto;
+pub mod sign;
+pub mod user;
+pub mod wechat;
