--- src/server/middlewares/logger.rs
@@ -1,1 +1,0 @@
-
